+++ CHANGELOG.md
@@ -10,6 +10,7 @@ All notable changes to `@homebridge-plugins/homebridge-matter` will be documente
 - add logging for all `MatterRequest` events
 - fix commands for lock devices
 - update plugin to support `@matter/main@0.16.x`
+- improvements to various device implementations
 
 ## v0.1.6 (2025-10-24)